--- src/components/FilmCard/index.js
@@ -1,1 +1,0 @@
-export * from './FilmCard';
