+++ src/components/FilmsList/FilmsList.js
@@ -1,10 +1,11 @@
 import React from 'react';
-import PropTypes from 'prop-types';
 import './FilmsList.scss';
-import { FilmCard } from '../FilmCard';
 
-export const FilmsList = (props) => {
-  const { films } = props;
+import { store } from '../../store/reducers';
+import { FilmCard } from '../FilmCard/FilmCard';
+
+export const FilmsList = () => {
+  const films = store.getState();
 
   return (
     <div className="films">
@@ -13,17 +14,4 @@ export const FilmsList = (props) => {
       ))}
     </div>
   );
-};
-
-FilmsList.propTypes = {
-  films: PropTypes.arrayOf(PropTypes.shape({
-    title: PropTypes.string.isRequired,
-    description: PropTypes.string,
-    imgUrl: PropTypes.string.isRequired,
-    imdbUrl: PropTypes.string.isRequired,
-  })),
-};
-
-FilmsList.defaultProps = {
-  films: [],
 };
