--- src/components/FormField/index.js
@@ -1,1 +1,0 @@
-export * from './FormField';
