--- src/components/NewFilm/index.js
@@ -1,1 +1,0 @@
-export * from './NewFilm';
