+++ src/store/action.js
@@ -1,0 +1,8 @@
+export const ACTION_TYPES = {
+  ADD_NEW_FILM: 'FILM::ADD',
+};
+
+export const addNewFilm = film => ({
+  type: ACTION_TYPES.ADD_NEW_FILM,
+  payload: film,
+});
